pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "langvel"
version = "0.1.0"
description = "A Laravel-inspired framework for building LangGraph agents with elegant developer experience"
authors = [
    {name = "Langvel Contributors"}
]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
keywords = ["langgraph", "ai", "agents", "framework", "laravel"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-anthropic>=0.2.0",
    "langchain-openai>=0.2.0",
    "pydantic>=2.0.0",
    "click>=8.0.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
    "chromadb>=0.4.0",
    "psycopg2-binary>=2.9.0",
    "redis>=5.0.0",
    "httpx>=0.27.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
langvel = "langvel.cli.main:cli"

[project.urls]
Homepage = "https://github.com/yourusername/langvel"
Documentation = "https://langvel.dev"
Repository = "https://github.com/yourusername/langvel"

[tool.setuptools.packages.find]
where = ["."]
include = ["langvel*"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
